+++ src/models/event.rs
@@ -8,7 +8,7 @@ pub struct Event {
     pub name: String,
     pub function: String, // String representation of the function to execute
     pub args: HashMap<String, EventArgs>,
-    pub timestamp: i32, // Event timestamp in ps
+    pub timestamp: i64, // Event timestamp in ps
 }
 
 impl Event {
@@ -16,7 +16,7 @@ impl Event {
         name: String,
         function: String,
         args: HashMap<String, EventArgs>,
-        timestamp: i32,
+        timestamp: i64,
     ) -> Event {
         Event {
             name,
